Removed old "Plug And Play" project from pluto's vs.net solution. Fixed small warning.

+++ src/DCE/DataGrid.cpp
@@ -158,7 +158,7 @@ DataGridCell::DataGridCell(int Size, char *Data)
 	}
 	if (m_NumAttributes)
 	{
-		for(int i=0;i<m_NumAttributes;++i)
+		for(size_t i=0;i<m_NumAttributes;++i)
 		{
 			char *pAttribute = Data;
 			Data += strlen(pAttribute) +1;
